requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
prefect==2.16.6
dbt-postgres==1.7.14
jupyter==1.0.0
matplotlib==3.8.4
